pyproject.toml
[project]
name = "content-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Zeynep Zeytun",email = "zeynepzeytun.2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "selenium (>=4.38.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]

[tool.poetry]
packages = [{include = "content_bot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "ruff (>=0.14.7,<0.15.0)"
]
